OT132-91: Move some method that use only one test.
